+++ run.py
@@ -6,6 +6,7 @@ import os
 # import flywheel functions
 from flywheel_gear_toolkit import GearToolkitContext
 from app.main import bet
+from app.parser import parse_config
 
 # The gear is split up into 2 main components. The run.py file which is executed
 # when the container runs. The run.py file then imports the rest of the gear as a
@@ -15,7 +16,8 @@ log = logging.getLogger(__name__)
 
 def main(context: GearToolkitContext) -> None:
     """Parses config and runs."""    
-    bet()
+    input_image = parse_config(context)
+    bet(input_image)
 
 # Only execute if file is run as main, not when imported by another module
 if __name__ == "__main__":  # pragma: no cover